requirements.txt
# ========================================
# DEPENDÊNCIAS DO PROJETO
# ========================================

# Framework principal
streamlit==1.31.0

# Manipulação de dados
pandas==2.1.4
numpy==1.26.3

# Visualizações
plotly==5.18.0

# Requisições HTTP
requests==2.31.0

# Configurações YAML
pyyaml==6.0.1

# Datas e horas
python-dateutil==2.8.2

# Utilitários
pathlib==1.0.1
